requirements.txt
flask==3.0.0
playwright==1.40.0
google-genai==0.2.2
pillow==10.1.0
requests==2.31.0
python-dotenv==1.0.0
beautifulsoup4==4.12.2
pandas